# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Village DTH Platform API
  version: 1.0.0
  description: |
    Production API contract for DTH plan purchase and service-request operations.
    Versioned via URI namespace `/api/v1`.
servers:
  - url: /api/v1
    description: Versioned API base path
...
security: []
paths:
  /auth/admin/login:
    post:
      tags: [Auth]
      operationId: adminLogin
      summary: Admin login and token issuance
      description: Authenticate admin and return access/refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
            examples:
              default:
                value:
                  identifier: "admin"
                  password: "StrongPassword123!"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
              examples:
                success:
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI..."
                    token_type: "Bearer"
                    expires_in: 864000
                    refresh_token: "rt_abc123"
                    refresh_expires_in: 7776000
                    user:
                      id: "usr_1"
                      name: "Village Admin"
                      phone_number: "9000000001"
                      role: "admin"
                      status: "active"
                      created_at: "2026-02-27T10:00:00Z"
                      updated_at: "2026-02-27T10:00:00Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /customer/checkout/service-requests:
    post:
      tags: [Customer, Payments]
      operationId: createServiceRequestAndPayment
      summary: Create service request and initiate payment
      description: Creates/uses customer record, creates service request, creates Razorpay order.
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
            examples:
              default:
                value:
                  provider_id: "pro_1"
                  plan_id: "pln_1"
                  customer:
                    name: "Kumar"
                    phone_number: "9000000002"
                  recaptcha_token: "03AFcWeA..."
      responses:
        "201":
          description: Service request and payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
              examples:
                created:
                  value:
                    service_request:
                      id: "sr_1"
                      user_id: "usr_2"
                      plan_id: "pln_1"
                      status: "pending"
                      created_at: "2026-02-27T10:10:00Z"
                      updated_at: "2026-02-27T10:10:00Z"
                    payment:
                      gateway: "razorpay"
                      order_id: "order_Qx1Ab2C3"
                      amount: 279
                      currency: "INR"
                      status: "created"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequestsError"
        "500":
          $ref: "#/components/responses/InternalError"
  /customer/service-requests/{serviceRequestId}/retry-payment:
    post:
      tags: [Customer, Payments]
      operationId: retryPaymentForServiceRequest
      summary: Retry payment using the same service request
      description: Reuses existing service request and creates a new payment attempt.
      security: []
      parameters:
        - $ref: "#/components/parameters/ServiceRequestIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recaptcha_token]
              properties:
                recaptcha_token:
                  type: string
                  minLength: 10
      responses:
        "200":
          description: Retry payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /payments/razorpay/callback:
    post:
      tags: [Payments]
      operationId: handleRazorpayCallback
      summary: Razorpay callback/webhook endpoint
      description: Verifies gateway signature and updates payment/service request status.
      security: []
      parameters:
        - name: X-Razorpay-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RazorpayCallbackRequest"
      responses:
        "200":
          description: Callback processed
          content:
            application/json:
              schema:
                type: object
                required: [acknowledged]
                properties:
                  acknowledged:
                    type: boolean
                    enum: [true]
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          description: Invalid gateway signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate callback safely ignored
          content:
            application/json:
              schema:
                type: object
                required: [acknowledged, duplicate]
                properties:
                  acknowledged:
                    type: boolean
                  duplicate:
                    type: boolean
        "500":
          $ref: "#/components/responses/InternalError"
...
  /admin/providers/csv-upload:
    post:
      tags: [Admin Providers]
      operationId: uploadProvidersCsv
      summary: Upsert providers from CSV
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: CSV processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvUploadResult"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /admin/plans/csv-upload:
    post:
      tags: [Admin Plans]
      operationId: uploadPlansCsv
      summary: Upsert plans from CSV
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: CSV processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvUploadResult"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /admin/service-requests/{serviceRequestId}/status:
    patch:
      tags: [Admin Service Requests]
      operationId: updateServiceRequestStatus
      summary: Update service request status
      description: Enforces allowed transitions; triggers SMS when transitioning to completed.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceRequestIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestStatusUpdateRequest"
            examples:
              toCompleted:
                value:
                  status: "completed"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ServiceRequestIdPath:
      name: serviceRequestId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: Unique key to prevent duplicate payment operations.
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictError:
      description: Conflict / duplicate / idempotency conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequestsError:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    User:
      type: object
      required: [id, name, phone_number, role, status, created_at, updated_at]
      properties:
        id:
          type: string
          example: "usr_1"
        name:
          type: string
          minLength: 1
          maxLength: 120
        phone_number:
          type: string
          pattern: "^[6-9][0-9]{9}$"
        role:
          type: string
          enum: [customer, admin]
        status:
          type: string
          enum: [active, inactive, blocked]
        password:
          type: string
          minLength: 8
          writeOnly: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ServiceRequestStatus:
      type: string
      enum: [pending, payment_failed, payment_success, blocked, completed]
      description: Canonical API values for service request state.
    ServiceRequest:
      type: object
      required: [id, user_id, plan_id, status, created_at, updated_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        plan_id:
          type: string
        status:
          $ref: "#/components/schemas/ServiceRequestStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaymentRequest:
      type: object
      required: [provider_id, plan_id, customer, recaptcha_token]
      properties:
        provider_id:
          type: string
        plan_id:
          type: string
        customer:
          type: object
          required: [name, phone_number]
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 120
            phone_number:
              type: string
              pattern: "^[6-9][0-9]{9}$"
        recaptcha_token:
          type: string
          minLength: 10
    PaymentResponse:
      type: object
      required: [service_request, payment]
      properties:
        service_request:
          $ref: "#/components/schemas/ServiceRequest"
        payment:
          type: object
          required: [gateway, order_id, amount, currency, status]
          properties:
            gateway:
              type: string
              enum: [razorpay]
            order_id:
              type: string
            amount:
              type: number
              minimum: 0.01
              multipleOf: 0.01
            currency:
              type: string
              enum: [INR]
            status:
              type: string
              enum: [created, authorized, captured, failed]
    RazorpayCallbackRequest:
      type: object
      required: [event, payload]
      properties:
        event:
          type: string
          example: "payment.captured"
        payload:
          type: object
          additionalProperties: true
...
    ServiceRequestStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ServiceRequestStatus"
    CsvUploadResult:
      type: object
      required: [total_rows, success_rows, failed_rows, errors]
      properties:
        total_rows:
          type: integer
          minimum: 0
        success_rows:
          type: integer
          minimum: 0
        failed_rows:
          type: integer
          minimum: 0
        errors:
          type: array
          items:
            $ref: "#/components/schemas/CsvRowError"
      example:
        total_rows: 3
        success_rows: 2
        failed_rows: 1
        errors:
          - row: 3
            field: "name"
            issue: "duplicate_plan_name_for_provider"
    CsvRowError:
      type: object
      required: [row, issue]
      properties:
        row:
          type: integer
          minimum: 1
        field:
          type: string
        issue:
          type: string
...
    AdminLoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 120
          description: Username or registered phone number
        password:
          type: string
          minLength: 8
    AdminLoginResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token, refresh_expires_in, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 864000
        refresh_token:
          type: string
        refresh_expires_in:
          type: integer
          description: Refresh token TTL in seconds
          example: 7776000
        user:
          $ref: "#/components/schemas/User"
    ErrorDetail:
      type: object
      required: [field, issue]
      properties:
        field:
          type: string
        issue:
          type: string
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"